HCI: Set p_gate to the gate used by the pipe and not to the first one
BZ: 140386

The connectivity pipe is sometimes deleted when restarting the NFC stack.
The issue is due to the fact that the owner of the connectivity gate
which should be set to 0000 is unexpectedly modified and set to 8000.

The function nfa_hciu_find_gate_by_gid should be used in order to
retreive the gate with a given id. If this function is properly
used then there is no need to change the owner of this gate.

Category: device enablement
Domain: CWS-NFC.NCI
Origin: internal
Upstream-Candidate: no, proprietary

Signed-off-by: Sebastien Griffoul <sebastien.griffoul@intel.com>
Change-Id: I703ba5070ff3c9bf93c966b536c80d1f34401b81
Reviewed-on: http://android.intel.com:8080/138644
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>