[PORT FROM MAIN JB-MR2]REVERT-ME  HCI: Create the conn. pipe if it doesn't exist
BZ: 140386

There are few use cases where we can have a mismatch on the HCI state
between the host terminal and the UICC. These use cases are the factory
reset and new build flashing (in both cases the /data partition content
is lost).

The root cause of the issue is that the content of /data/nfc (in which
the host terminal stores the HCI network state) is lost. If the UICC has
opened the connectivity pipe between the host terminal and itself, the
UICC may believe that this pipe is still opened whereas the host
terminal doesn't know about it. This prevents all the connectivity
events from being correctly handled.

Workaround is to manually re-create this pipe when such events are
received. This is possible because the host controller always assigns
the same id number to these kind of pipes. So the host terminal can
deduce all the useful pipe parameters only from the pipe id.

The final solution should consists in sending a ADM_CLEAR_ALL_PIPES
command when device detects that there is a mismatch on the SESSION_ID
parameters. Unfortunately it looks like the ADM_CLEAR_ALL_PIPES is
not correctly broadcasted to the UICC.

Change-Id: I8abd750b02f01904fa1cecfe05a11d62e36dffa7
Orig-Change-Id: I65faf5179ee2d285e89f9d792fe828ae9b256fcc
Category: device enablement
Domain: CWS.NFC-NCI
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Sebastien Griffoul <sebastien.griffoul@intel.com>
Reviewed-on: http://android.intel.com:8080/138941
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>